staymap/requirements.txt
flet
flet-map
sqlmodel
psycopg2-binary
pydantic>=1.10.0
pydantic-settings>=0.1.0
python-dotenv